refactor: 更新日志模块，重构日志导入路径并移除旧的logger文件
