+++ src/massflow/module/mass_spectrum_set.py
@@ -11,7 +11,7 @@ License: See LICENSE file in project root
 """
 from typing import List, Tuple,Literal,overload,Union, Optional
 import matplotlib.pyplot as plt
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 from massflow.module.ms_meta_data import MetaDataBase,ImzMlMetaData
 from massflow.module.spectrum import Spectrum
