+++ src/massflow/module/ms_data_manager_imzml.py
@@ -14,7 +14,7 @@ import warnings
 import numpy as np
 from pyimzml.metadata import ParamGroup
 from pyimzml.ImzMLParser import ImzMLParser
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 from massflow.module.ms_data_manager import MSDataManager
 from massflow.module.spectrum_imzml import SpectrumImzML
 from massflow.module.ms_meta_data import ImzMlMetaData
