+++ src/massflow/r_preprocess/adapter.py
@@ -1,7 +1,7 @@
 from massflow.module.mass_spectrum_set import MassSpectrumSet
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
 from massflow.r_preprocess.environment import REnvironment
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 
 logger = get_logger("r_preprocess_adapter")
