+++ src/massflow/tools/__init__.py
@@ -1,0 +1,7 @@
+"""Utilities for MassFlow.
+
+Note: This package coexists with a legacy `__init___.py` file in the same
+folder. Use `massflow.tools` imports going forward.
+"""
+
+from .logger import get_logger  # re-export for convenience
