+++ tests/test_all_baseline.py
@@ -6,7 +6,7 @@ import pytest
 from massflow.module.mass_spectrum_set import MassSpectrumSet
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
 from massflow.preprocess.spectrum_preprocess import SpectrumPreprocess
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 
 logger = get_logger("test_all_baseline")
