+++ tests/test_noise_reduction.py
@@ -4,7 +4,7 @@ from functools import partial
 import pytest
 from massflow.module.mass_spectrum_set import MassSpectrumSet
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 from massflow.preprocess.spectrum_preprocess import SpectrumPreprocess
 
 logger = get_logger("test_noise_reduction")
