+++ tests/test_peak_alignment.py
@@ -2,7 +2,7 @@ import pytest
 import time
 import numpy as np
 from typing import Optional
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 from massflow.module.mass_spectrum_set import MassSpectrumSet
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
 from massflow.preprocess.spectrum_preprocess import SpectrumPreprocess
