+++ tests/test_read.py
@@ -2,7 +2,7 @@ import time
 import subprocess
 import pytest
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 
 logger = get_logger("test_read")
