+++ tests/test_swap.py
@@ -4,7 +4,7 @@ from functools import partial
 import pytest
 from massflow.module.spectrum_imzml import SpectrumImzML
 from massflow.module.ms_data_manager_imzml import MSDataManagerImzML
-from massflow.logger import get_logger
+from massflow.tools.logger import get_logger
 
 logger = get_logger("test_swap")
